removed file with test_ at start for the test client
