+++ todo_app/tests/test_trello_pages.py
@@ -1,7 +1,21 @@
 
-from todo_app.test_client import client
 import requests
 import os
+import pytest
+from dotenv import load_dotenv, find_dotenv
+from todo_app import app
+
+@pytest.fixture
+def client():
+    # Use our test integration config instead of the 'real' version
+    file_path = find_dotenv('../.env.test')
+    load_dotenv(file_path, override=True)
+
+    # Create the new app.
+    test_app = app.create_app()
+
+    with test_app.test_client() as client:
+        yield client
 
 class StubResponse():
